References Oppevara/oppevara#67, editing is fully functional.
It could always be made better, but the initial functionality is
working. It could be beneficial to fully synchronise the data and the
DOM. This would mean that textual values from the inputs should be
updated into the data elements once changes occur. This would require to
initialize and keep udpdated the column textual values for a single row.
The benefit would be that storage would not require to pick data up from
the DOM.