Added linter and fixed any outlined issues.
Added linter with configuration that is suggested by the H5P team. Installed all the required modules and added package.json and package-lock.json files to the repository. Added .gitignore file and updated the .h5pignore to ignore all the newly added files that should be ignored by the packager. Added a script that will run the linter that can be ran with "npm run lint".